Add null/wrong-type guard clauses in processor
Be more defensive in RedactQueryStringValueTelemetryProcessor
about the type of objects that are being processed.

- Only process RequestTelemetry
- Better handling of null values for the options object
- Only process absolute URIs (relative URIs are not touched)

Adds more tests around passing in the other telemetry type
objects, things with nulls, etc.

Relative URLs may be handled in a future version.  Our immediate
need is for absolute URL argument values in the Url.Query property.