Add filtering for DependencyTelemetry QueryStrings
Adds the capability to also filter out QueryString argument
values for DependencyTelemetry objects in Azure Application
Insights.  It uses the same options object as what is used for
the RequestTelemetry processing.

There are two new booleans on the options object which
can disable the redaction for either ITelemetry type. The
default for both is 'true'.